requirements.txt
Flask==2.3.2
flask-cors==4.0.0
google-generativeai==0.5.4
chromadb==0.4.24
sentence-transformers==2.7.0
numpy==1.26.4
huggingface_hub==0.24.7
tokenizers>=0.19,<0.20
transformers==4.44.2
chromadb==0.4.24
